requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.37.18
botocore==1.37.18
celery==5.3.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-admin-interface==0.27.0
django-colorfield==0.12.0
django-cors-headers==4.3.1
django-filter==23.3
django-ses==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
jmespath==1.0.1
kombu==5.5.0
packaging==24.2
Pillow==10.0.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2025.1
redis==5.0.1
reportlab==4.0.7
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
